project/requirements.txt
asyncio==3.4.3
attrs==20.3.0
autopep8==2.3.1
backports.shutil-get-terminal-size==1.0.0
backports.shutil_which==3.5.2
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.9.0
braceexpand==0.1.7
Brotli==1.1.0
cachelib==0.13.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
check50==3.3.11
clang-format==19.1.5
cli50==8.0.1
click==8.1.8
colorama==0.4.6
compare50==1.2.6
contourpy==1.3.1
cryptography==44.0.0
cs50==9.4.0
CS50-VSIX-Client @ file:///opt/cs50/extensions/cs50vsix-client
cssselect2==0.7.0
cycler==0.12.1
djhtml==3.0.7
EditorConfig==0.17.0
Flask==3.1.0
Flask-Session==0.8.0
fonttools==4.55.3
greenlet==3.1.1
help50==3.0.5
icdiff==2.0.7
idna==3.10
inflect==7.0.0
iniconfig==2.0.0
intervaltree==3.1.0
itsdangerous==2.2.0
jellyfish==0.11.2
Jinja2==3.1.5
jsbeautifier==1.15.1
kiwisolver==1.4.8
lib50==3.0.12
MarkupSafe==3.0.2
matplotlib==3.10.0
more-itertools==10.5.0
msgspec==0.19.0
mypy-extensions==1.0.0
natsort==8.4.0
numpy==1.26.4
packaging==24.2
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
psutil==6.1.1
ptyprocess==0.7.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==1.10.21
pydyf==0.11.0
Pygments==2.18.0
pyparsing==3.2.1
pypdf==5.1.0
pyphen==0.17.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
render50==9.2.7
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.3
style50==2.10.4
submit50==3.2.0
tabulate==0.9.0
termcolor==1.1.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typeguard==4.4.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
weasyprint==63.1
webencodings==0.5.1
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
zopfli==0.2.3.post1
